+++ src/main/java/org/terasology/basiccrafting/PatternRecipe.java
@@ -1,0 +1,12 @@
+// Copyright 2020 The Terasology Foundation
+// SPDX-License-Identifier: Apache-2.0
+package org.terasology.basiccrafting;
+
+import org.terasology.basiccrafting.components.Recipe;
+
+/**
+ * Test second recipe type
+ */
+public class PatternRecipe implements Recipe {
+
+}
