+++ src/main/java/org/terasology/basiccrafting/components/Recipe.java
@@ -1,0 +1,9 @@
+// Copyright 2020 The Terasology Foundation
+// SPDX-License-Identifier: Apache-2.0
+package org.terasology.basiccrafting.components;
+
+/**
+ * The interface to represent a base recipe type.
+ */
+public interface Recipe {
+}
