+++ src/main/java/org/terasology/basiccrafting/events/OnRecipeCrafted.java
@@ -1,0 +1,23 @@
+// Copyright 2020 The Terasology Foundation
+// SPDX-License-Identifier: Apache-2.0
+
+package org.terasology.basiccrafting.events;
+
+import org.terasology.engine.entitySystem.entity.EntityRef;
+import org.terasology.engine.entitySystem.event.Event;
+
+/**
+ * This event is sent on an item when it is crafted in a workstation. The event is sent on the produce with the
+ * ingredients as parameters. Can be used to modify the produce with some values dictated by the ingredients.
+ */
+public class OnRecipeCrafted implements Event {
+    private final EntityRef[] ingredients;
+
+    public OnRecipeCrafted(EntityRef[] ingredients) {
+        this.ingredients = ingredients;
+    }
+
+    public EntityRef[] getIngredients() {
+        return ingredients;
+    }
+}
