+++ src/main/java/org/terasology/basiccrafting/listCrafting/components/CraftingWorkstationComponent.java
@@ -1,0 +1,9 @@
+// Copyright 2020 The Terasology Foundation
+// SPDX-License-Identifier: Apache-2.0
+package org.terasology.basiccrafting.listCrafting.components;
+
+import org.terasology.engine.entitySystem.Component;
+
+public class CraftingWorkstationComponent implements Component {
+    public String recipeCategory;
+}
